refactor : Entity 분리
확장 가능성을 염두하고, 모든 Entity를 고유한 폴더로 분리함
